+++ bhp/lims/skins/bhp_scripts/guard_submit.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2018 Botswana Harvard Partnership (BHP)
+
+## Script (Python) "guard_submit"
+##bind container=container
+##bind context=context
+##bind namespace=
+##bind script=script
+##bind subpath=traverse_subpath
+##parameters=
+##title=
+##
+
+from bhp.lims.workflow.guards import guard_submit
+return guard_submit(context)