+++ bhp/lims/workflow/guards.py
@@ -2,12 +2,16 @@
 #
 # Copyright 2018 Botswana Harvard Partnership (BHP)
 
+from bhp.lims import logger
 from bika.lims import api
 from bika.lims.interfaces import IAnalysisRequest, ISamplePartition, ISample
 
 # This is required to authorize AccessControl.ZopeGuards to access to this
 # module and function/s via skin's python scripts.
 from AccessControl.SecurityInfo import ModuleSecurityInfo
+from bika.lims.workflow import isBasicTransitionAllowed, isActive, \
+    getCurrentState, wasTransitionPerformed
+
 security = ModuleSecurityInfo(__name__)
 
 @security.public
@@ -88,6 +92,50 @@ def guard_send_to_pot(context):
     """Guard for sending the sample to the point of testing
     """
     return api.is_active(context)
+
+
+@security.public
+def guard_submit(context):
+    if not IAnalysisRequest.providedBy(context):
+        # Note that this guard is only used for bika_ar_workflow!
+        return True
+
+    logger.info("*** Custom Guard: submit **")
+    if not isBasicTransitionAllowed(context):
+        return False
+
+    invalid = 0
+    analyses = context.getAnalyses()
+    for an in analyses:
+        if an.review_state == 'to_be_verified':
+            continue
+
+        # The analysis has already been verified?
+        an = api.get_object(an)
+        if wasTransitionPerformed(an, 'submit'):
+            continue
+
+        # Maybe the analysis is in an 'inactive' state?
+        if not isActive(an):
+            invalid += 1
+            continue
+
+        # Maybe the analysis has been rejected or retracted?
+        dettached = ['rejected', 'retracted', 'attachments_due']
+        status = getCurrentState(an)
+        if status in dettached:
+            invalid += 1
+            continue
+
+        # At this point we can assume this analysis is an a valid state and
+        # the AR could potentially be submitted, but the Analysis Request can
+        # only be submitted if all the analyses have been submitted already
+        return False
+
+    # Be sure that at least there is one analysis in an active state, it
+    # doesn't make sense to submit an Analysis Request if all the analyses that
+    # contains are rejected or cancelled!
+    return len(analyses) - invalid > 0
 
 
 def get_sample(instance):